Refactoring: group some code by modules
